requirements.txt
aiosqlite==0.22.1
asyncpg==0.31.0
psycopg2==2.9.10
pydantic==2.12.5
python-dotenv==1.2.1
SQLAlchemy==2.0.46
streamlit==1.50.0
temporalio==1.18.2
# protobuf pinned to be compatible with all dependencies
protobuf>=3.20,<6.0.0
google-generativeai>=0.8.0
llama-index-core>=0.11.0
llama-index-llms-google-genai>=0.2.0
# Search & Extraction
perplexityai
tavily-python
llama-cloud>=1.1.0
trafilatura
beautifulsoup4
crawl4ai>=0.7.4
pillow
fastapi
uvicorn
python-multipart
